pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[project]
name = "seo-agent"
version = "0.1.0"
description = "Autonomous SEO agent for keyword research and recommendations"
authors = [
    {name = "Viachaslau Kazakou", email = "sly.kazakoff@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = "^3.11"
keywords = ["seo", "agent", "nlp", "keywords"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.24.0",
    "httpx>=0.25.0",
    "playwright>=1.40.0",
    "trafilatura>=1.6.0",
    "readability-lxml>=0.8.1",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "scikit-learn>=1.3.0",
    "sentence-transformers>=2.2.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "click>=8.1.0",
    "sqlalchemy>=2.0.0",
    "sqlmodel>=0.0.14",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "isort>=5.13.0",
]

[project.scripts]
seo-agent = "seo_agent.cli.main:cli"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
lint.select = ["E", "F", "W", "I"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.isort]
profile = "black"
line_length = 100
